llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "mcp>=1.1.2",
    "ollama>=0.4.4",
    "requests>=2.32.3",
    "hydra-core>=1.2.0",
    "python-dotenv>=1.0.1",
    "chainlit>=2.2.1",
    "aiohttp>=3.11.12",
]
